power operator

+++ cMatrix.cpp
@@ -696,7 +696,7 @@ CMatrix CMatrix::Log(const CMatrix &a )
 
     return temp ;
 }
-CMatrix  CMatrix::operator ^=( const int a)
+CMatrix  CMatrix::operator ^=( const double a)
 {
     CMatrix temp (this->nR , this->nC);
     for (int i=0 ; i<this->nR ;i++ )
